Refactor integration tests package structure for better modularity
