+++ integration-tests/src/jvmTest/kotlin/ai/jetbrains/code/integration/tests/MultipleLLMPromptExecutorIntegrationTest.kt
@@ -1,8 +1,10 @@
 package ai.jetbrains.code.integration.tests
 
-import ai.jetbrains.code.integration.tests.TestUtils.readTestAnthropicKeyFromEnv
-import ai.jetbrains.code.integration.tests.TestUtils.readTestGoogleAIKeyFromEnv
-import ai.jetbrains.code.integration.tests.TestUtils.readTestOpenAIKeyFromEnv
+import ai.jetbrains.code.integration.tests.utils.Models
+import ai.jetbrains.code.integration.tests.utils.TestUtils
+import ai.jetbrains.code.integration.tests.utils.TestUtils.readTestAnthropicKeyFromEnv
+import ai.jetbrains.code.integration.tests.utils.TestUtils.readTestGoogleAIKeyFromEnv
+import ai.jetbrains.code.integration.tests.utils.TestUtils.readTestOpenAIKeyFromEnv
 import ai.koog.agents.core.tools.ToolDescriptor
 import ai.koog.agents.core.tools.ToolParameterDescriptor
 import ai.koog.agents.core.tools.ToolParameterType
