+++ integration-tests/src/jvmTest/kotlin/ai/jetbrains/code/integration/tests/OpenRouterLLMClientIntegrationTest.kt
@@ -1,9 +1,11 @@
-package ai.koog.prompt.executor.clients.openrouter
+package ai.jetbrains.code.integration.tests
 
 import ai.koog.agents.core.tools.ToolDescriptor
 import ai.koog.agents.core.tools.ToolParameterDescriptor
 import ai.koog.agents.core.tools.ToolParameterType
 import ai.koog.prompt.dsl.Prompt
+import ai.koog.prompt.executor.clients.openrouter.OpenRouterLLMClient
+import ai.koog.prompt.executor.clients.openrouter.OpenRouterModels
 import ai.koog.prompt.message.Message
 import kotlinx.coroutines.flow.toList
 import kotlinx.coroutines.test.runTest
