+++ integration-tests/src/jvmTest/kotlin/ai/jetbrains/code/integration/tests/utils/Models.kt
@@ -1,4 +1,4 @@
-package ai.jetbrains.code.integration.tests
+package ai.jetbrains.code.integration.tests.utils
 
 import ai.koog.prompt.executor.clients.anthropic.AnthropicModels
 import ai.koog.prompt.executor.clients.google.GoogleModels
