+++ integration-tests/src/jvmTest/kotlin/ai/jetbrains/code/integration/tests/utils/TestLogger.kt
@@ -1,4 +1,4 @@
-package ai.jetbrains.code.integration.tests
+package ai.jetbrains.code.integration.tests.utils
 
 import ai.koog.agents.local.features.common.message.FeatureMessage
 import ai.koog.agents.local.features.common.message.FeatureMessageProcessor
